Code review response, add nothing to be normalized error
When a harvester gathers raw documents, but does not normalize any documents because
none of them are in the approved set_spec list, the linter would raise an index error
when trying to check if the element was serializable. With this change, the approprite
value error is raised if there's nothing to normalize because of the set_spec check.