Add openaire harvester, as OAI PMH for now
In the future, we want to investigate many more sources for data as
OpenAire has lots of rich APIs

Using the OAI-PMH endpoint is a quick way to start using it as a data
source right away. We can explore more routes of collaboration as we go
forward.

Addresses [#SHARE-14] and closes #274